ITS#8890 Handle sizeof(time_t) > sizeof(long) in format strings
64-bit time_t means that on some architectures, time_t is now larger
than a long, and making some references in format strings incorrect.
To avoid truncation or other size mismatch issues, always cast to a long
long and read using %lld.

+++ contrib/slapd-modules/smbk5pwd/smbk5pwd.c
@@ -510,10 +510,10 @@ static int smbk5pwd_exop_passwd(
 		qpw->rs_mods = ml;
 
 		keys = ch_malloc( 2 * sizeof(struct berval) );
-		keys[0].bv_val = ch_malloc( LDAP_PVT_INTTYPE_CHARS(long) );
+		keys[0].bv_val = ch_malloc( LDAP_PVT_INTTYPE_CHARS(long long) );
 		keys[0].bv_len = snprintf(keys[0].bv_val,
-			LDAP_PVT_INTTYPE_CHARS(long),
-			"%ld", slap_get_time());
+			LDAP_PVT_INTTYPE_CHARS(long long),
+			"%lld", (long long)slap_get_time());
 		BER_BVZERO( &keys[1] );
 		
 		ml->sml_desc = ad_sambaPwdLastSet;
@@ -532,10 +532,10 @@ static int smbk5pwd_exop_passwd(
 			qpw->rs_mods = ml;
 
 			keys = ch_malloc( 2 * sizeof(struct berval) );
-			keys[0].bv_val = ch_malloc( LDAP_PVT_INTTYPE_CHARS(long) );
+			keys[0].bv_val = ch_malloc( LDAP_PVT_INTTYPE_CHARS(long long) );
 			keys[0].bv_len = snprintf(keys[0].bv_val,
-					LDAP_PVT_INTTYPE_CHARS(long),
-					"%ld", slap_get_time() + pi->smb_must_change);
+					LDAP_PVT_INTTYPE_CHARS(long long),
+					"%lld", (long long)(slap_get_time() + pi->smb_must_change));
 			BER_BVZERO( &keys[1] );
 
 			ml->sml_desc = ad_sambaPwdMustChange;
@@ -555,10 +555,10 @@ static int smbk5pwd_exop_passwd(
 			qpw->rs_mods = ml;
 
 			keys = ch_malloc( 2 * sizeof(struct berval) );
-			keys[0].bv_val = ch_malloc( LDAP_PVT_INTTYPE_CHARS(long) );
+			keys[0].bv_val = ch_malloc( LDAP_PVT_INTTYPE_CHARS(long long) );
 			keys[0].bv_len = snprintf(keys[0].bv_val,
-					LDAP_PVT_INTTYPE_CHARS(long),
-					"%ld", slap_get_time() + pi->smb_can_change);
+					LDAP_PVT_INTTYPE_CHARS(long long),
+					"%lld", (long long)(slap_get_time() + pi->smb_can_change));
 			BER_BVZERO( &keys[1] );
 
 			ml->sml_desc = ad_sambaPwdCanChange;
