+++ libraries/libldap/os-local.c
@@ -164,8 +164,8 @@ ldap_pvt_connect(LDAP *ld, ber_socket_t s, struct sockaddr_un *sa, int async)
 	}
 
 	Debug3(LDAP_DEBUG_TRACE,
-		"ldap_connect_timeout: fd: %d tm: %ld async: %d\n",
-		s, opt_tv ? tv.tv_sec : -1L, async);
+		"ldap_connect_timeout: fd: %d tm: %lld async: %d\n",
+		s, (long long)(opt_tv ? tv.tv_sec : -1L), async);
 
 	if ( ldap_pvt_ndelay_on(ld, s) == -1 ) return -1;
