+++ servers/slapd/back-asyncmeta/meta_result.c
@@ -1723,7 +1723,7 @@ void* asyncmeta_timeout_loop(void *ctx, void *arg)
 	LDAP_STAILQ_INIT( &timeout_list );
 	a_metasingleconn_t* oldest[mi->mi_ntargets];
 
-	Debug( asyncmeta_debug, "asyncmeta_timeout_loop[%p] start at [%ld] \n", rtask, current_time );
+	Debug( asyncmeta_debug, "asyncmeta_timeout_loop[%p] start at [%lld] \n", rtask, (long long)current_time );
 	if ( mi->mi_disabled > 0 && asyncmeta_db_has_pending_ops( mi ) == 0 ) {
 		Debug( asyncmeta_debug, "asyncmeta_timeout_loop[%p] database disabled, clearing connections [%ld] \n", rtask, current_time );
 		ldap_pvt_thread_mutex_lock( &mi->mi_mc_mutex );
@@ -1844,9 +1844,9 @@ void* asyncmeta_timeout_loop(void *ctx, void *arg)
 				a_metasingleconn_t *log_msc =  &mc->mc_conns[0];
 				Debug( asyncmeta_debug,
 				       "asyncmeta_timeout_loop:Timeout op %s loop[%p], "
-				       "current_time:%ld, op->o_time:%ld msc: %p, "
+				       "current_time:%lld, op->o_time:%lld msc: %p, "
 				       "msc->msc_binding_time: %x, msc->msc_flags:%x \n",
-				       bc->op->o_log_prefix, rtask, current_time, bc->op->o_time,
+				       bc->op->o_log_prefix, rtask, (long long)current_time, (long long)bc->op->o_time,
 				       log_msc, (unsigned int)log_msc->msc_binding_time, log_msc->msc_mscflags );
 
 				if (bc->searchtime) {
@@ -1934,7 +1934,7 @@ void* asyncmeta_timeout_loop(void *ctx, void *arg)
 	}
 	slap_sl_mem_setctx(ctx, oldctx);
 	current_time = slap_get_time();
-	Debug( asyncmeta_debug, "asyncmeta_timeout_loop[%p] stop at [%ld] \n", rtask, current_time );
+	Debug( asyncmeta_debug, "asyncmeta_timeout_loop[%p] stop at [%lld] \n", rtask, (long long)current_time );
 	ldap_pvt_thread_mutex_lock( &slapd_rq.rq_mutex );
 	if ( ldap_pvt_runqueue_isrunning( &slapd_rq, rtask )) {
 		ldap_pvt_runqueue_stoptask( &slapd_rq, rtask );
